Restored warning component file so it'll be easier to delete later
