+++ frontend/packages/core/src/Feedback/index.tsx
@@ -3,6 +3,7 @@ export { default as Error } from "./error";
 export { default as Hint } from "./hint";
 export { Note, NotePanel } from "./note";
 export { Tooltip, TooltipContainer } from "./tooltip";
-export { default as Toast, Warning } from "./toast";
+export { default as Toast } from "./toast";
+export { default as Warning } from "./warning";
 
 export type { NoteConfig } from "./note";
